pyproject.toml
[project]
name = "line-bot"
version = "0.1.0"
description = "Line Bot application"
requires-python = ">=3.9"
dependencies = [
    "line-bot-sdk",
    "flask",
    "google-generativeai",
    "requests",
    "python-dotenv",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
